requirements.txt
pandas==2.2.3
openai==1.108.2
llama-index==0.14.2
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-huggingface==0.6.1
pdfminer.six==20250506
torch==2.8.0
torchvision==0.23.0
python-dotenv==1.1.1
pymilvus==2.6.2
tqdm==4.67.1
llama-index-vector-stores-milvus==0.9.2
milvus-lite==2.5.1
mxbai-rerank==0.1.6